requirements.txt
# Core
playwright>=1.40.0
playwright-stealth>=1.0.6
apscheduler>=3.10.0
pyyaml>=6.0
typer>=0.9.0
python-dotenv>=1.0.0

# AI
openai>=1.0.0
httpx>=0.25.0

# Image Processing
Pillow>=10.0.0

# CLI Output
rich>=13.0.0


# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
